+++ .travis.yml
@@ -25,9 +25,9 @@ addons:
             - dvipng
 python:
     - 2.7
-    # - 3.3
-    # - 3.4
-    # - 3.5
+    - 3.3
+    - 3.4
+    - 3.5
 env:
     global:
         # The following versions are the 'default' for tests, unless
