+++ bin/generate_model
@@ -1,5 +1,5 @@
 #!/usr/bin/env python
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 from sys import exit
-from desidatamodel.stub.main import main
+from desidatamodel.stub import main
 exit(main())
