+++ py/desidatamodel/test/generate_fits.py
@@ -12,6 +12,7 @@ def generate_fits():
     from ..stub import process_file
     image = np.random.random_integers(0,2**16-1,size=(100,100)).astype(np.uint16)
     hdu0 = fits.PrimaryHDU(image,uint=True)
+    hdu0.header.comments['BZERO'] = 'Data are really unsigned 16-bit int.'
     a1 = np.array(['NGC1001', 'NGC1002', 'NGC1003'])
     a2 = np.array([11.1, 12.3, 15.2],dtype='f')
     a3 = np.array([57.6, 99.43, 123.4])
@@ -20,6 +21,7 @@ def generate_fits():
     col3 = fits.Column(name='vdisp', format='D', array=a3, unit='km/s')
     cols = fits.ColDefs([col1, col2, col3])
     hdu1 = fits.BinTableHDU.from_columns(cols)
+    hdu1.header['EXTNAME'] = ('Galaxies', 'Name of this HDU.')
     hdulist = fits.HDUList([hdu0,hdu1])
     filename = join(dirname(__file__),'t','fits_file.fits')
     if exists(filename):
